+++ npm/ng-packs/packages/account-core/proxy/src/lib/proxy/account/index.ts
@@ -1,4 +1,5 @@
 import * as Web from './web';
 export * from './account.service';
 export * from './models';
+export * from './profile.service';
 export { Web };
