+++ npm/ng-packs/packages/account-core/proxy/src/lib/proxy/account/models.ts
@@ -1,5 +1,21 @@
 import type { ExtensibleObject } from '@abp/ng.core';
 
+export interface ChangePasswordInput {
+  currentPassword?: string;
+  newPassword: string;
+}
+
+export interface ProfileDto extends ExtensibleObject {
+  userName?: string;
+  email?: string;
+  name?: string;
+  surname?: string;
+  phoneNumber?: string;
+  isExternal: boolean;
+  hasPassword: boolean;
+  concurrencyStamp?: string;
+}
+
 export interface RegisterDto extends ExtensibleObject {
   userName: string;
   emailAddress: string;
@@ -18,4 +34,13 @@ export interface SendPasswordResetCodeDto {
   appName: string;
   returnUrl?: string;
   returnUrlHash?: string;
+}
+
+export interface UpdateProfileDto extends ExtensibleObject {
+  userName?: string;
+  email?: string;
+  name?: string;
+  surname?: string;
+  phoneNumber?: string;
+  concurrencyStamp?: string;
 }
