+++ npm/ng-packs/packages/account/package.json
@@ -8,7 +8,6 @@
   },
   "dependencies": {
     "@abp/ng.account.core": "~5.0.0-beta.2",
-    "@abp/ng.identity": "~5.0.0-beta.2",
     "@abp/ng.theme.shared": "~5.0.0-beta.2",
     "tslib": "^2.0.0"
   },
