+++ npm/ng-packs/packages/account/src/lib/components/manage-profile/manage-profile.component.ts
@@ -1,4 +1,4 @@
-import { ProfileService } from '@abp/ng.identity/proxy';
+import { ProfileService } from '@abp/ng.account.core/proxy';
 import { fadeIn } from '@abp/ng.theme.shared';
 import { transition, trigger, useAnimation } from '@angular/animations';
 import { Component, OnInit } from '@angular/core';
