+++ npm/ng-packs/packages/account/src/lib/components/personal-settings/personal-settings.component.ts
@@ -1,4 +1,4 @@
-import { ProfileService } from '@abp/ng.identity/proxy';
+import { ProfileService } from '@abp/ng.account.core/proxy';
 import { ToasterService } from '@abp/ng.theme.shared';
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
