+++ npm/ng-packs/packages/identity/proxy/src/lib/proxy/identity/index.ts
@@ -2,4 +2,3 @@ export * from './identity-role.service';
 export * from './identity-user-lookup.service';
 export * from './identity-user.service';
 export * from './models';
-export * from './profile.service';
