+++ npm/ng-packs/packages/identity/proxy/src/lib/proxy/identity/models.ts
@@ -1,9 +1,9 @@
-import type { ExtensibleEntityDto, ExtensibleFullAuditedEntityDto, ExtensibleObject, PagedAndSortedResultRequestDto } from '@abp/ng.core';
-
-export interface ChangePasswordInput {
-  currentPassword?: string;
-  newPassword: string;
-}
+import type {
+  ExtensibleEntityDto,
+  ExtensibleFullAuditedEntityDto,
+  ExtensibleObject,
+  PagedAndSortedResultRequestDto,
+} from '@abp/ng.core';
 
 export interface GetIdentityRolesInput extends PagedAndSortedResultRequestDto {
   filter?: string;
@@ -13,8 +13,7 @@ export interface GetIdentityUsersInput extends PagedAndSortedResultRequestDto {
   filter?: string;
 }
 
-export interface IdentityRoleCreateDto extends IdentityRoleCreateOrUpdateDtoBase {
-}
+export interface IdentityRoleCreateDto extends IdentityRoleCreateOrUpdateDtoBase {}
 
 export interface IdentityRoleCreateOrUpdateDtoBase extends ExtensibleObject {
   name: string;
@@ -71,26 +70,6 @@ export interface IdentityUserUpdateDto extends IdentityUserCreateOrUpdateDtoBase
 
 export interface IdentityUserUpdateRolesDto {
   roleNames: string[];
-}
-
-export interface ProfileDto extends ExtensibleObject {
-  userName?: string;
-  email?: string;
-  name?: string;
-  surname?: string;
-  phoneNumber?: string;
-  isExternal: boolean;
-  hasPassword: boolean;
-  concurrencyStamp?: string;
-}
-
-export interface UpdateProfileDto extends ExtensibleObject {
-  userName?: string;
-  email?: string;
-  name?: string;
-  surname?: string;
-  phoneNumber?: string;
-  concurrencyStamp?: string;
 }
 
 export interface UserLookupCountInputDto {
